requirements.txt
Django==4.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
boto3==1.37.35
botocore==1.37.35
celery==5.5.1
pillow==11.2.1
redis==5.2.1
locust==1.0.3
django[gis]
django-redis>=5.0.0